Revert "Team summary"
