+++ src/Gambler.java
@@ -8,13 +8,13 @@ public class Gambler extends Hero {
 	/**
 	 * A brief string description of the hero's special ability.
 	 */
-	private static final String SPECIAL_ABILITY = "Advantages in villain battles.";
+	private static final String SPECIAL_ABILITY = "Advantages in villain battles";
 	
 	/**
 	 * An extended string description of the hero.
 	 * This description is shown to the player when selecting heroes for their team.
 	 */
-	private static final String DESCRIPTION = "You don't have to look very far to find a gambler these days, but true experts are few and far between.\nThese players use more than luck to tip the odds in their favour.\n\nWhile fighting villains, Gamblers gain the following advantages:\nPaper Scissors Rock: Reveal one option that the villain will not play.\nGuess the Number: Get an extra guess.\nDice Roll: Increase your roll by one.";
+	private static final String DESCRIPTION = "You don't have to look very far to find a gambler these days, but true experts are few and far between. These players use more than luck to tip the odds in their favour.\nWhile fighting villains, Gamblers gain the following advantages:\nPaper Scissors Rock: Reveal one option that the villain will not play.\nGuess the Number: Get an extra guess.\nDice Roll: Increase your roll by one.\n";
 	
 	/**
 	 * The maximum health the hero can have.
