--- src/SetUpPanel.java
@@ -1,533 +1,0 @@
-import java.awt.Dimension;
-import java.awt.EventQueue;
-
-import javax.swing.JFrame;
-import javax.swing.JPanel;
-import javax.swing.JLabel;
-import javax.swing.BoxLayout;
-import javax.swing.SwingConstants;
-import java.awt.Font;
-import javax.swing.JButton;
-import java.awt.event.ActionListener;
-import java.awt.event.ActionEvent;
-import javax.swing.ImageIcon;
-import java.awt.Color;
-import java.awt.Rectangle;
-import java.awt.CardLayout;
-import javax.swing.JTextField;
-import javax.swing.JComboBox;
-import javax.swing.DefaultComboBoxModel;
-import java.awt.Component;
-import java.awt.Container;
-
-import javax.swing.border.LineBorder;
-import javax.swing.JTextPane;
-import javax.swing.UIManager;
-import java.awt.GridLayout;
-
-public class SetUpPanel extends JPanel {
-	
-	/**
-	 * A string representation of this panel, used by the CardLayout in Game.
-	 */
-	public static final String SET_UP_PANEL_STRING = "Set Up Panel";
-
-	/**
-	 * A string representation of the start screen panel, used by the CardLayout
-	 * of the main content panel.
-	 */
-	private static final String START_SCREEN_PANEL_STRING = "Start Screen Panel";
-
-	/**
-	 * A string representation of the create team panel, used by the CardLayout
-	 * of the main content panel.
-	 */
-	private static final String CREATE_TEAM_PANEL_STRING = "Create Team Panel";
-
-	/**
-	 * A string representation of the add hero panel, used by the CardLayout
-	 * of the main content panel.
-	 */
-	private static final String ADD_HERO_PANEL_STRING = "Add Hero Panel";
-	
-	/**
-	 * A string representation of the team summary panel, used by the CardLayout
-	 * of the main content panel.
-	 */
-	private static final String TEAM_SUMMARY_PANEL_STRING = "Team Summary Panel";
-	
-	/**
-	 * The main Game window that this panel is a part of. Used by event handlers
-	 * to change the panel shown in game.
-	 */
-	private Game gameWindow;
-	
-	/**
-	 * The main sub-panel of the SetUpPanel, located below the game title.
-	 * Uses a CardLayout to hold several other panels.
-	 */
-	private JPanel contentPanel;
-	
-	/**
-	 * A CardLayout applied to the main content panel, used to hold several
-	 * other panels.
-	 */
-	private CardLayout contentPanelCardLayout;
-	
-	/**
-	 * A panel for the start screen, containing a "new game" button and an "exit" button.
-	 */
-	private JPanel startScreenPanel;
-
-	/**
-	 * A panel which contains components to ask the user for a team name,
-	 * number of heroes, and number of cities.
-	 */
-	private JPanel createTeamPanel;
-
-	/**
-	 * A panel which contains components to ask the user to select a name
-	 * and type for a hero, and display information about the different hero types.
-	 */
-	private JPanel addHeroPanel;
-	
-	/**
-	 * A panel which contains components to display a summary of the newly created
-	 * team to the user, and allow them to begin the game.
-	 */
-	private JPanel teamSummaryPanel;
-	
-	// Add Hero panel components
-	private JLabel lblCreateHero;
-	private JPanel heroInfoPanel;
-	private CardLayout heroInfoPanelCardLayout;
-	private JPanel inputPanel;
-	private JLabel lblInvalidNameErrorMessage;
-	private JTextField txtfName;
-
-	// Team Summary panel components
-	private JPanel heroSummariesPanel;	
-	
-	/**
-	 * Constructor for a SetUpPanel. Creates a JPanel of the appropriate size,
-	 * and adds all the required components.
-	 * @param gameWindow		The main game window which this panel is a part of.
-	 */
-	public SetUpPanel(Game gameWindow) {
-		super();
-		setBounds(new Rectangle(0, 0, 880, 610));
-		this.gameWindow = gameWindow;
-		setLayout(null);
-		
-		addTitle();
-		addContentPanel();
-		addSubPanels();
-	}
-	
-	/**
-	 * Helper method to return the team of the current GameEnvironment.
-	 * @return		The team of the current GameEnvirionment.
-	 */
-	private Team getTeam() {
-		return gameWindow.getGame().getTeam();
-	}
-	
-	/**
-	 * Creates a label containing the game's title and adds it to the panel.
-	 */
-	private void addTitle() {
-		JLabel lblTitle = new JLabel("HEROES AND VILLAINS");
-		lblTitle.setFont(new Font("Tahoma", Font.PLAIN, 80));
-		lblTitle.setHorizontalAlignment(SwingConstants.CENTER);
-		lblTitle.setBounds(10, 11, 860, 150);
-		add(lblTitle);		
-		
-	}
-
-	/**
-	 * Creates a JPanel below the title to hold the main content of the
-	 * set up page, and gives it a CardLayout to hold the different panels
-	 * shown during game set up.
-	 */
-	private void addContentPanel() {
-		contentPanel = new JPanel();
-		contentPanel.setBounds(10, 172, 860, 427);
-		add(contentPanel);
-		
-		contentPanelCardLayout = new CardLayout();
-		contentPanel.setLayout(contentPanelCardLayout);
-		
-	}
-	
-	/**
-	 * Creates a JPanel for every sub-panel within the main content panel,
-	 * and adds them to the CardLayout.
-	 * Sets the start screen panel to be show initially.
-	 */
-	private void addSubPanels() {
-		startScreenPanel = new JPanel();
-		initialiseStartScreenPanel();
-		contentPanel.add(startScreenPanel, START_SCREEN_PANEL_STRING);
-
-		createTeamPanel = new JPanel();
-		initialiseCreateTeamPanel();
-		contentPanel.add(createTeamPanel, CREATE_TEAM_PANEL_STRING);
-		
-		addHeroPanel = new JPanel();
-		initialiseAddHeroPanel();
-		contentPanel.add(addHeroPanel, ADD_HERO_PANEL_STRING);
-		
-		initialiseTeamSummaryPanel();
-		
-		contentPanelCardLayout.show(contentPanel, START_SCREEN_PANEL_STRING);
-		
-	}
-	
-	
-	
-	
-	
-	/* 
-	 * -------------------------------------------------------------------------------------
-	 * Start Screen Panel
-	 * -------------------------------------------------------------------------------------
-	 */
-
-	/**
-	 * Adds components to the start screen panel.
-	 */
-	private void initialiseStartScreenPanel() {
-		startScreenPanel.setLayout(null);
-		
-		JButton btnNewGame = new JButton("New Game");
-		btnNewGame.setFont(new Font("Tahoma", Font.PLAIN, 20));
-		btnNewGame.addActionListener(new ActionListener() {
-			public void actionPerformed(ActionEvent e) {
-				contentPanelCardLayout.show(contentPanel, CREATE_TEAM_PANEL_STRING);
-			}
-		});
-		btnNewGame.setBounds(346, 116, 168, 57);
-		startScreenPanel.add(btnNewGame);
-		
-		JButton btnExit = new JButton("Exit");
-		btnExit.addActionListener(new ActionListener() {
-			public void actionPerformed(ActionEvent e) {
-				System.exit(0);
-			}
-		});
-		btnExit.setFont(new Font("Tahoma", Font.PLAIN, 20));
-		btnExit.setBounds(346, 256, 168, 57);
-		startScreenPanel.add(btnExit);
-	}
-	
-	
-	
-	
-	/* 
-	 * -------------------------------------------------------------------------------------
-	 * Create Team Panel
-	 * -------------------------------------------------------------------------------------
-	 */
-	
-	/**
-	 * Add components to the create team panel.
-	 */
-	private void initialiseCreateTeamPanel() {
-		createTeamPanel.setLayout(null);
-		
-		JLabel lblEnterTeamName = new JLabel("Enter a name for you team of heroes:");
-		lblEnterTeamName.setHorizontalAlignment(SwingConstants.TRAILING);
-		lblEnterTeamName.setFont(new Font("Tahoma", Font.PLAIN, 16));
-		lblEnterTeamName.setBounds(10, 74, 414, 32);
-		createTeamPanel.add(lblEnterTeamName);
-		
-		JTextField txtfTeamName = new JTextField();
-		txtfTeamName.setBounds(447, 74, 222, 32);
-		createTeamPanel.add(txtfTeamName);
-		txtfTeamName.setColumns(10);
-		
-		JLabel lblTeamNameErrorMessage = new JLabel("Team name must be between 2 and 10 characters long.");
-		lblTeamNameErrorMessage.setVisible(false);
-		lblTeamNameErrorMessage.setBounds(447, 113, 403, 14);
-		createTeamPanel.add(lblTeamNameErrorMessage);
-		
-		JLabel lblChooseNumHeroes = new JLabel("How many heroes would you like on your team?");
-		lblChooseNumHeroes.setHorizontalAlignment(SwingConstants.TRAILING);
-		lblChooseNumHeroes.setFont(new Font("Tahoma", Font.PLAIN, 16));
-		lblChooseNumHeroes.setBounds(10, 150, 414, 32);
-		createTeamPanel.add(lblChooseNumHeroes);
-		
-		JComboBox<Integer> cmbNumHeroes = new JComboBox<Integer>();
-		cmbNumHeroes.setModel(new DefaultComboBoxModel<Integer>(new Integer[] {1, 2, 3}));
-		cmbNumHeroes.setFont(new Font("Tahoma", Font.PLAIN, 16));
-		cmbNumHeroes.setBounds(447, 150, 66, 32);
-		createTeamPanel.add(cmbNumHeroes);
-		
-		JLabel lblChooseNumCities = new JLabel("How many cities would you like to play?");
-		lblChooseNumCities.setHorizontalAlignment(SwingConstants.TRAILING);
-		lblChooseNumCities.setFont(new Font("Tahoma", Font.PLAIN, 16));
-		lblChooseNumCities.setBounds(10, 230, 414, 32);
-		createTeamPanel.add(lblChooseNumCities);
-		
-		JComboBox<Integer> cmbNumCities = new JComboBox<Integer>();
-		cmbNumCities.setModel(new DefaultComboBoxModel<Integer>(new Integer[] {3, 4, 5, 6}));
-		cmbNumCities.setFont(new Font("Tahoma", Font.PLAIN, 16));
-		cmbNumCities.setBounds(447, 230, 66, 32);
-		createTeamPanel.add(cmbNumCities);
-		
-		JButton btnCreateTeamNext = new JButton("Next");
-		btnCreateTeamNext.addActionListener(new ActionListener() {
-			public void actionPerformed(ActionEvent e) {
-				String teamName = txtfTeamName.getText();
-				int numHeroes = (int) cmbNumHeroes.getSelectedItem();
-				int numCities = (int) cmbNumCities.getSelectedItem();
-				
-				if (Team.isValidTeamName(teamName)) {
-					gameWindow.setGame(new GameEnvironment(teamName, numHeroes, numCities));
-					refreshAddHeroPanel();
-					contentPanelCardLayout.show(contentPanel, ADD_HERO_PANEL_STRING);
-				} else {
-					lblTeamNameErrorMessage.setVisible(true);
-				}
-			}
-		});
-		btnCreateTeamNext.setFont(new Font("Tahoma", Font.PLAIN, 16));
-		btnCreateTeamNext.setBounds(381, 320, 98, 32);
-		createTeamPanel.add(btnCreateTeamNext);
-	}
-	
-	
-	
-	
-	
-	/* 
-	 * -------------------------------------------------------------------------------------
-	 * Add Hero Panel
-	 * -------------------------------------------------------------------------------------
-	 */
-	
-	/**
-	 * Adds components to the add hero panel.
-	 */
-	private void initialiseAddHeroPanel() {
-		addHeroPanel.setLayout(null);
-		
-		inputPanel = new JPanel();
-		inputPanel.setBounds(10, 11, 300, 405);
-		addHeroPanel.add(inputPanel);
-		
-		initialiseInputPanel();
-		
-		heroInfoPanel = new JPanel();
-		heroInfoPanel.setBorder(new LineBorder(new Color(0, 0, 0)));
-		heroInfoPanel.setBounds(320, 11, 530, 405);
-		addHeroPanel.add(heroInfoPanel);
-		
-		heroInfoPanelCardLayout = new CardLayout();
-		heroInfoPanel.setLayout(heroInfoPanelCardLayout);
-		
-		addHeroInfoPanels();		
-
-	}
-	
-	private void initialiseInputPanel() {
-		inputPanel.setLayout(null);
-		
-		lblCreateHero = new JLabel("Create Hero");
-		lblCreateHero.setBounds(10, 11, 280, 29);
-		lblCreateHero.setHorizontalAlignment(SwingConstants.CENTER);
-		lblCreateHero.setFont(new Font("Tahoma", Font.PLAIN, 24));
-		inputPanel.add(lblCreateHero);
-		
-		JLabel lblEnterAName = new JLabel("Enter a name:");
-		lblEnterAName.setBounds(20, 73, 260, 20);
-		lblEnterAName.setFont(new Font("Tahoma", Font.PLAIN, 16));
-		inputPanel.add(lblEnterAName);
-		
-		txtfName = new JTextField();
-		txtfName.setFont(new Font("Tahoma", Font.PLAIN, 16));
-		txtfName.setBounds(30, 104, 240, 29);
-		txtfName.setColumns(10);
-		inputPanel.add(txtfName);
-		
-		lblInvalidNameErrorMessage = new JLabel("");
-		lblInvalidNameErrorMessage.setBounds(30, 142, 250, 14);
-		inputPanel.add(lblInvalidNameErrorMessage);
-		
-		JLabel lblSelectHeroType = new JLabel("Select a hero type:");
-		lblSelectHeroType.setBounds(20, 177, 260, 20);
-		lblSelectHeroType.setFont(new Font("Tahoma", Font.PLAIN, 16));
-		inputPanel.add(lblSelectHeroType);
-		
-		JComboBox<String> cmbHeroTypes = new JComboBox<String>();
-		cmbHeroTypes.addActionListener(new ActionListener() {
-			public void actionPerformed(ActionEvent e) {
-				heroInfoPanelCardLayout.show(heroInfoPanel,
-										(String) cmbHeroTypes.getSelectedItem());
-			}
-		});
-		cmbHeroTypes.setFont(new Font("Tahoma", Font.PLAIN, 16));
-		cmbHeroTypes.setModel(new DefaultComboBoxModel<String>(Hero.allHeroTypes()));
-		cmbHeroTypes.setBounds(30, 208, 240, 29);
-		inputPanel.add(cmbHeroTypes);
-		
-		JButton btnAddHeroToTeam = new JButton("Add Hero to Team");
-		btnAddHeroToTeam.addActionListener(new ActionListener() {
-			public void actionPerformed(ActionEvent e) {
-				
-				String heroName = txtfName.getText();
-				String type = (String) cmbHeroTypes.getSelectedItem();
-				
-				if (! getTeam().isValidName(heroName)) {
-					lblInvalidNameErrorMessage.setText("That name is already taken by another hero.");
-					
-				} else if (heroName.trim().isEmpty()) {
-					lblInvalidNameErrorMessage.setText("Invalid name.");
-									
-				} else {
-					getTeam().addHero(heroName, type);
-					if (getTeam().getHeroes().size() < getTeam().getStartNumHeroes()) {
-						refreshAddHeroPanel();
-					} else {
-						contentPanelCardLayout.show(contentPanel, TEAM_SUMMARY_PANEL_STRING);
-						addHeroSummaries();
-					}
-				}
-			}
-		});
-		btnAddHeroToTeam.setFont(new Font("Tahoma", Font.PLAIN, 16));
-		btnAddHeroToTeam.setBounds(43, 335, 214, 29);
-		inputPanel.add(btnAddHeroToTeam);
-		
-	}
-
-	/**
-	 * Adds an information panel for each hero to the hero info panel
-	 * card layout, using the hero's type as the String reference.
-	 */
-	private void addHeroInfoPanels() {
-		for (Hero hero: Hero.ALL_HEROES) {
-			heroInfoPanel.add(makeHeroInfoPanel(hero), hero.getType());
-		}
-		
-	}
-
-	/**
-	 * Takes a hero subclass instance, and creates an information panel for
-	 * a hero of that type, including its image, type, special ability,
-	 * max health and description.
-	 * @param hero		An instance of the hero subclass which the information
-	 * 					panel is describing.
-	 * @return			The hero information JPanel.
-	 */
-	private JPanel makeHeroInfoPanel(Hero hero) {
-		JPanel infoPanel = new JPanel();
-		infoPanel.setLayout(null);
-		
-		JLabel heroImage = new JLabel("");
-		heroImage.setBounds(10, 11, 200, 200);
-		infoPanel.add(heroImage);
-		
-		JLabel lblType = new JLabel(hero.getType());
-		lblType.setFont(new Font("Tahoma", Font.PLAIN, 24));
-		lblType.setHorizontalAlignment(SwingConstants.CENTER);
-		lblType.setBounds(220, 11, 298, 37);
-		infoPanel.add(lblType);
-		
-		JTextPane txtpnDescription = new JTextPane();
-		txtpnDescription.setBackground(UIManager.getColor("Panel.background"));
-		txtpnDescription.setFont(new Font("Tahoma", Font.PLAIN, 13));
-		txtpnDescription.setText(hero.getDescription());
-		txtpnDescription.setBounds(10, 222, 508, 170);
-		infoPanel.add(txtpnDescription);
-		
-		JLabel lblSpecialAbility = new JLabel("Special Ability:");
-		lblSpecialAbility.setFont(new Font("Tahoma", Font.PLAIN, 16));
-		lblSpecialAbility.setBounds(220, 59, 298, 20);
-		infoPanel.add(lblSpecialAbility);
-		
-		JLabel lblHerosSpecialAbility = new JLabel(hero.getSpecialAbility());
-		lblHerosSpecialAbility.setFont(new Font("Tahoma", Font.PLAIN, 16));
-		lblHerosSpecialAbility.setBounds(230, 90, 288, 20);
-		infoPanel.add(lblHerosSpecialAbility);
-		
-		JLabel lblMaxHealth = new JLabel(String.format("Max Health: %s", hero.getMaxHealth()));
-		lblMaxHealth.setFont(new Font("Tahoma", Font.PLAIN, 16));
-		lblMaxHealth.setBounds(220, 131, 298, 20);
-		infoPanel.add(lblMaxHealth);
-		
-		return infoPanel;
-	}
-
-	/**
-	 * Called after each hero is added. Updates the add hero panel by
-	 * updating the hero number, emptying the name text field, and
-	 * resetting the selected hero type the the default.
-	 */
-	private void refreshAddHeroPanel() {
-		lblCreateHero.setText(String.format("Create Hero %s:",
-				getTeam().getHeroes().size() + 1));
-		
-		txtfName.setText("");
-		
-		lblInvalidNameErrorMessage.setText("");
-
-	}
-
-	
-	
-	
-	
-	/* 
-	 * -------------------------------------------------------------------------------------
-	 * Team Summary Panel
-	 * -------------------------------------------------------------------------------------
-	 */
-	
-	private void initialiseTeamSummaryPanel() {
-		teamSummaryPanel = new JPanel();
-		contentPanel.add(teamSummaryPanel, TEAM_SUMMARY_PANEL_STRING);
-		teamSummaryPanel.setLayout(null);
-		
-		JLabel lblTeamName = new JLabel("Team Name");
-		lblTeamName.setFont(new Font("Tahoma", Font.PLAIN, 24));
-		lblTeamName.setHorizontalAlignment(SwingConstants.CENTER);
-		lblTeamName.setBounds(10, 11, 840, 70);
-		teamSummaryPanel.add(lblTeamName);
-		
-		JButton btnStartGame = new JButton("Start Game");
-		btnStartGame.setFont(new Font("Tahoma", Font.PLAIN, 16));
-		btnStartGame.setBounds(369, 367, 122, 36);
-		teamSummaryPanel.add(btnStartGame);
-		
-		heroSummariesPanel = new JPanel();
-		heroSummariesPanel.setBounds(10, 92, 840, 262);
-		teamSummaryPanel.add(heroSummariesPanel);
-	}
-	
-	/**
-	 * Adds the hero summaries to the team summary panel.
-	 * Called once the team has been created.
-	 */
-	private void addHeroSummaries() {
-		GridLayout teamSummaryGridLayout = new GridLayout(0, getTeam().getHeroes().size(), 0, 0);
-		heroSummariesPanel.setLayout(teamSummaryGridLayout);
-		
-		for (Hero hero: getTeam().getHeroes()) {
-			heroSummariesPanel.add(heroSummaryPanel(hero));
-		}
-	}
-	
-	/**
-	 * Takes a hero, and produces a panel containing the hero's image,
-	 * name and type.
-	 * @param hero		The hero for which to make a summary panel.
-	 * @return			A summary panel for the given hero.
-	 */
-	private JPanel heroSummaryPanel(Hero hero) {
-		return new JPanel();
-	}
-	
-	
-	
-}
