+++ src/Team.java
@@ -27,11 +27,6 @@ public class Team {
 	private ArrayList<Hero> heroes = new ArrayList<Hero>();
 	
 	/**
-	 * The number of heroes to be added to the team initially.
-	 */
-	private int startNumHeroes;
-	
-	/**
 	 * A String representing the name of the team.
 	 */
 	private String name;
@@ -57,28 +52,6 @@ public class Team {
 	public Team(String name) {
 		this.name = name;
 		this.currentMoney = STARTING_MONEY;
-	}
-	
-	/**
-	 * A constructor for team, which sets the starting number of
-	 * heroes which need to be added to the team.
-	 * @param name				The team's name.
-	 * @param startNumHeroes	The starting number of heroes of the team.
-	 */
-	public Team(String name, int startNumHeroes) {
-		this.name = name;
-		this.startNumHeroes = startNumHeroes;
-		this.currentMoney = STARTING_MONEY;
-	}
-	
-	/**
-	 * Checks whether the given team name is valid. A name is valid
-	 * if it is between 2 and 10 characters long.
-	 * @param name		The team name to check.
-	 * @return			true if the name is valid, false otherwise.
-	 */
-	public static boolean isValidTeamName(String name) {
-		return name.length() >= 2 && name.length() <= 10;
 	}
 
 	/**
@@ -111,31 +84,20 @@ public class Team {
 		
 		String type = Hero.ALL_HEROES[heroIndex].getType();
 		
-		addHero(heroName, type);
-
-	}
-	
-	/**
-	 * Creates a new hero with the given name of the given type,
-	 * and adds it to the team's list of heroes.
-	 * @param name	The name of the new hero.
-	 * @param type	The type of the new hero.
-	 */
-	public void addHero(String heroName, String type) {
 		switch (type) {
 			case "Apprentice": heroes.add(new Apprentice(heroName));
-						   	   break;
+							   break;
 			case "Bulwark":    heroes.add(new Bulwark(heroName));
 							   break;
 			case "Explorer":   heroes.add(new Explorer(heroName));
-		                   	   break;
+			                   break;
 			case "Gambler":    heroes.add(new Gambler(heroName));
-		                       break;
+			                   break;
 			case "Mercenary":  heroes.add(new Mercenary(heroName));
-		                       break;
+			                   break;
 			case "Merchant":   heroes.add(new Merchant(heroName));
-						       break;
-			default:           throw new RuntimeException("No such hero type");
+							   break;
+		    default:           throw new RuntimeException("No such hero type");
 		}
 	}
 	
@@ -145,7 +107,7 @@ public class Team {
 	 * on the team.
 	 * @return	true if the name is valid, false otherwise.
 	 */
-	public boolean isValidName(String heroName) {
+	private boolean isValidName(String heroName) {
 		boolean isValid  = true;
 		for (Hero hero: heroes) {
 			if (hero.getName().equals(heroName)) {
@@ -351,10 +313,6 @@ public class Team {
 		return heroes;
 	}
 	
-	public int getStartNumHeroes() {
-		return startNumHeroes;
-	}
-
 	/**
 	 * Getter method for numMaps.
 	 * @return The value of numMaps.
